Fixed compiler warning
git-svn-id: http://miranda.googlecode.com/svn/trunk@12061 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/utils/utf.cpp
@@ -99,7 +99,7 @@ int Ucs2toUtf8Len(const wchar_t *src)
 	if ( src == 0 )
 		return 0;
 
-	return Ucs2toUtf8Len( src, wcslen( src ));
+	return Ucs2toUtf8Len( src, (int)wcslen( src ));
 }
 
 /* wide char to UTF-8 string conversion */
